+++ go/api/glide_client.go
@@ -382,3 +382,18 @@ func (client *GlideClient) LastSave() (int64, error) {
 	}
 	return handleIntResponse(response)
 }
+
+// Resets the statistics reported by the server using the INFO and LATENCY HISTOGRAM.
+//
+// Return value:
+//
+//	OK to confirm that the statistics were successfully reset.
+//
+// [valkey.io]: https://valkey.io/commands/config-resetstat/
+func (client *GlideClient) ConfigResetStat() (string, error) {
+	response, err := client.executeCommand(C.ConfigResetStat, []string{})
+	if err != nil {
+		return DefaultStringResponse, err
+	}
+	return handleStringResponse(response)
+}