+++ go/api/server_management_cluster_commands_test.go
@@ -218,3 +218,26 @@ func ExampleGlideClusterClient_LastSaveWithOptions() {
 
 	// Output: true
 }
+
+func ExampleGlideClusterClient_ConfigResetStat() {
+	var client *GlideClusterClient = getExampleGlideClusterClient() // example helper function
+	result, err := client.ConfigResetStat()
+	if err != nil {
+		fmt.Println("Glide example failed with an error: ", err)
+	}
+	fmt.Println(result)
+
+	// Output: OK
+}
+
+func ExampleGlideClusterClient_ConfigResetStatWithOptions() {
+	var client *GlideClusterClient = getExampleGlideClusterClient() // example helper function
+	opts := options.RouteOption{Route: nil}
+	result, err := client.ConfigResetStatWithOptions(opts)
+	if err != nil {
+		fmt.Println("Glide example failed with an error: ", err)
+	}
+	fmt.Println(result)
+
+	// Output: OK
+}