+++ go/api/server_management_commands_test.go
@@ -214,3 +214,15 @@ func ExampleGlideClient_LastSave() {
 
 	// Output: true
 }
+
+func ExampleGlideClient_ConfigResetStat() {
+	var client *GlideClient = getExampleGlideClient() // example helper function
+	response, err := client.ConfigResetStat()
+	if err != nil {
+		fmt.Println("Glide example failed with an error: ", err)
+	}
+	fmt.Println(response)
+
+	// Output:
+	// OK
+}